requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
discord.py==1.5.1
dnspython==2.0.0
docopt==0.6.2
idna==2.10
multidict==4.7.6
phonetics==1.0.5
pipreqs==0.4.10
pymongo==3.11.2
pytz==2020.4
requests==2.25.0
urllib3==1.26.2
wincertstore==0.2
yarg==0.1.9
yarl==1.5.1
